+++ libsolidity/formal/SymbolicTypes.cpp
@@ -672,12 +672,14 @@ std::map<std::string, frontend::Type const*> transactionMemberTypes()
 	// TODO: gasleft
 	return {
 		{"block.basefee", TypeProvider::uint256()},
+		{"block.blobbasefee", TypeProvider::uint256()},
 		{"block.chainid", TypeProvider::uint256()},
 		{"block.coinbase", TypeProvider::address()},
 		{"block.prevrandao", TypeProvider::uint256()},
 		{"block.gaslimit", TypeProvider::uint256()},
 		{"block.number", TypeProvider::uint256()},
 		{"block.timestamp", TypeProvider::uint256()},
+		{"blobhash", TypeProvider::array(DataLocation::Memory, TypeProvider::uint256())},
 		{"blockhash", TypeProvider::array(DataLocation::Memory, TypeProvider::uint256())},
 		{"msg.data", TypeProvider::bytesCalldata()},
 		{"msg.sender", TypeProvider::address()},
@@ -690,10 +692,10 @@ std::map<std::string, frontend::Type const*> transactionMemberTypes()
 
 std::map<std::string, SortPointer> transactionMemberSorts()
 {
-	// NOTE: `blockhash` needs proper `ArraySort`, `smtSort` wraps array types into array+length pair
+	// NOTE: `blockhash` and `blobhash` need proper `ArraySort`, `smtSort()` wraps array types into array+length pair
 	auto toSort = [&](auto const& entry) -> SortPointer
 	{
-		if (entry.first == "blockhash")
+		if (entry.first == "blockhash" || entry.first == "blobhash")
 			return std::make_shared<ArraySort>(SortProvider::uintSort, SortProvider::uintSort);
 		return smtSort(*entry.second);
 	};